whoops i made a mistake

+++ src/main/java/io/github/Gabriel/damagePlugin/customDamage/CustomDamager.java
@@ -18,7 +18,7 @@ public class CustomDamager {
         this.plugin = plugin;
     }
 
-    public void doCustomDamage(LivingEntity target, LivingEntity damager, DamageType type, double damage, DamageSourceType source) {
+    public void doCustomDamage(LivingEntity target, LivingEntity damager, DamageType type, int damage, DamageSourceType source) {
         UUID uuid = target.getUniqueId();
 
         // Only fire your custom damage event once here, directly
